Linear search

+++ 0x1E-search_algorithms/0-linear.c
@@ -4,24 +4,23 @@
  * linear_search - searches for value in int
  * array using Linear search algorithm
  *
- * @array: Input int array
- * @size: Array size
- * @value: Search value
- *
+ * @array: pointer to the first element of the array
+ * @size: number of elements in array
+ * @value: value to search for
  * Return: Always EXIT_SUCCESS
  */
 int linear_search(int *array, size_t size, int value)
 {
-	int i;
+	size_t i;
 
 	if (array == NULL)
 	{
 		return (-1);
 	}
 
-	for (i = 0; i < (int)size; i++)
+	for (i = 0; i < size; i++)
 	{
-		printf("Value checked array[%u] = [%d]\n", i, array[i]);
+		printf("Value checked array[%li] = [%i]\n", i, array[i]);
 		if (value == array[i])
 		{
 			return (i);
